refact: 빈 생성 순서 결정하는 로직 분리
