+++ nextstep-di/src/main/java/nextstep/di/beandefinition/BeanDefinitionRegistry.java
@@ -1,9 +1,12 @@
 package nextstep.di.beandefinition;
 
 import com.google.common.collect.Sets;
+import nextstep.di.exception.MultipleBeanImplementationException;
+import nextstep.di.exception.NotExistBeanException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -29,6 +32,25 @@ public class BeanDefinitionRegistry {
         return definitions.stream()
                 .filter(definition -> type.isAssignableFrom(definition.getBeanType()))
                 .collect(Collectors.toSet());
+    }
+
+    public BeanDefinition findExactBeanDefinition(Class<?> type) {
+        Set<BeanDefinition> definitions = findByType(type);
+
+        if (definitions.isEmpty()) {
+            throw NotExistBeanException.from(type);
+        }
+
+        if (2 <= definitions.size()) {
+            List<Class<?>> candidateTypes = definitions.stream()
+                    .map(definition -> definition.getBeanType())
+                    .collect(Collectors.toList());
+            throw MultipleBeanImplementationException.from(type, candidateTypes);
+        }
+
+        return definitions.stream()
+                .findFirst()
+                .get();
     }
 
     public Set<BeanDefinition> findAll() {
